Some small adjustements for Goodisplay

+++ applications_user/nfc_eink/nfc_eink_screen/goodisplay/eink_goodisplay_i.h
@@ -125,7 +125,7 @@ typedef struct {
     uint8_t update_cnt;
     uint8_t response_cnt;
 
-    uint8_t block_number;
+    uint16_t block_number;
     size_t data_index;
 } NfcEinkScreenSpecificGoodisplayContext;
